# backend/h2os/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
  /api/pools/{id}:
    get:
      tags:
      - pool-rest-controller
      summary: Get a pool by its ID.
      operationId: getPool
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: "Pool not found, probably invalid id supplied"
        "200":
          description: Pool found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoolDTO'
    put:
      tags:
      - pool-rest-controller
      summary: Edit a pool by its ID.
      operationId: editPool
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PoolDTO'
        required: true
      responses:
        "404":
          description: "Pool not found, probably invalid id supplied"
        "400":
          description: "Bad request, invalid data in changes"
        "401":
          description: "You are not authorized, you are not the admin"
        "200":
          description: Pool edited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoolDTO'
    delete:
      tags:
      - pool-rest-controller
      summary: Delete a pool by its ID.
      operationId: deletePool
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: "Pool not found, probably invalid id supplied"
        "401":
          description: "You are not authorized, you are not the admin"
        "204":
          description: Pool deleted
...
  /api/pools:
    get:
      tags:
      - pool-rest-controller
      summary: Get paged pools.
      operationId: getPools
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 5
      responses:
        "204":
          description: "Pools not found, probably high page number supplied"
        "200":
          description: Pools found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoolDTO'
    post:
      tags:
      - pool-rest-controller
      summary: Create a new pool.
      operationId: createPool
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PoolDTO'
        required: true
      responses:
        "401":
          description: "You are not authorized, you are not the admin"
        "201":
          description: Pool created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoolDTO'
        "400":
          description: "Bad request, invalid data in the body"
...
  /api/pools/{id}/messages:
    get:
      tags:
      - pool-rest-controller
      summary: Get all messages of a pool by its ID.
      operationId: getMessages
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 3
      responses:
        "200":
          description: Messages found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageDTO'
        "204":
          description: "Messages not found, probably high page number supplied"
        "404":
          description: "Pool not found, probably invalid id supplied"
    post:
      tags:
      - pool-rest-controller
      summary: Add a new message to a pool.
      operationId: addMessage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageDTO'
        required: true
      responses:
        "201":
          description: Message added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageDTO'
        "404":
          description: "Pool not found, probably invalid id supplied"
        "401":
          description: "You are not authorized, you are not the admin"
        "400":
          description: "Bad request, invalid data in the body"
...
components:
  schemas:
    PoolDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        direction:
          type: string
        capacity:
          type: integer
          format: int32
        scheduleStart:
          type: string
        scheduleEnd:
          type: string
        company:
          type: string
        description:
          type: string
    MessageDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        author:
          type: string
        body:
          type: string
...